backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
statsmodels>=0.14.0
APScheduler>=3.10.0
vaderSentiment>=3.3.2
gunicorn>=22.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
responses>=0.23.0
tenacity>=8.0.0
circuitbreaker>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
celery>=5.2.0
prometheus-client>=0.14.0
sentry-sdk>=1.9.0
plotly>=5.0.0
reportlab>=4.0.0
aiohttp>=3.8.0
flask-limiter>=3.5.0
marshmallow>=3.20.0
cryptography>=41.0.0
python-keycloak>=3.0.0
redis>=5.0.0
celery>=5.3.0
prometheus-client>=0.17.0
sentry-sdk[flask]>=1.40.0
psycopg2-binary>=2.9.0
sqlalchemy-utils>=0.41.0
websockets>=12.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
kafka-python>=2.0.0
pandas-profiling>=3.6.0
great-expectations>=0.17.0
mlflow>=2.8.0
shap>=0.43.0
scikit-learn>=1.3.0
plotly>=5.17.0
bokeh>=3.2.0
dash>=2.14.0
streamlit>=1.28.0
